Extracted subtract.
